requirements.txt
requests==2.31.0
python-dotenv==1.0.1
finance-calendars==0.0.3
pandas>=2.0.0
